Broken commit where intmax_t doesn't work yet

+++ src/nbr/ft_get_digit_count.c
@@ -6,7 +6,7 @@
 /*   By: sbos <sbos@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/10/20 12:00:48 by sbos          #+#    #+#                 */
-/*   Updated: 2022/02/04 17:38:13 by sbos          ########   odam.nl         */
+/*   Updated: 2022/02/05 16:39:01 by sbos          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
  * @param nbr
  * @return
  */
-size_t	ft_get_digit_count(int nbr)
+size_t	ft_get_digit_count(intmax_t nbr)
 {
 	return (ft_get_digit_count_base(nbr, 10));
 }
