+++ src/nbr/ft_nbr_to_str.c
@@ -11,8 +11,9 @@
 /* ************************************************************************** */
 
 #include "libft.h"
+#include "libft_nbr.h"
 
-static void	add_digits_to_str(char *str, int nbr, size_t number_len,
+static void	add_digits_to_str(char *str, intmax_t nbr, size_t number_len,
 								unsigned int base)
 {
 	int			digit;
@@ -30,14 +31,14 @@ static void	add_digits_to_str(char *str, int nbr, size_t number_len,
 }
 
 /**
- * @brief Converts signed int @p nbr to a string in @p base.
+ * @brief Converts intmax_t @p nbr to a string in @p base.
  * This function was made for ft_printf() its hex printing.
  *
  * @param number
  * @param base
  * @return
  */
-char	*ft_nbr_to_str(int number, unsigned int base)
+char	*ft_nbr_to_str(intmax_t number, unsigned int base)
 {
 	char			*str;
 	const size_t	number_len = ft_get_digit_count_base(number, base);
